register done

+++ index.php
@@ -1,5 +1,7 @@
 <?php
 include './utils/db.php';
+session_start();
+
 ?>
 <!doctype html>
 <html lang="en">
@@ -14,26 +16,61 @@ include './utils/db.php';
 </head>
 
 <body>
-<nav class="navbar navbar-expand-lg bg-body-tertiary">
-  <div class="container-fluid">
-    <a class="navbar-brand" href="/jeux">Games</a>
-    <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
-      <span class="navbar-toggler-icon"></span>
-    </button>
-    <div class="collapse navbar-collapse" id="navbarSupportedContent">
-      <ul class="navbar-nav me-auto mb-2 mb-lg-0">
-        <li class="nav-item">
-          <a class="nav-link active" aria-current="page" href="/jeux">Home</a>
-        </li>
-        <li class="nav-item">
-          <a class="nav-link" href="/jeux/new-game.php">Ajouter un jeu</a>
-        </li>
-        
-      </ul>
-      
+  <nav class="navbar navbar-expand-lg bg-body-tertiary">
+    <div class="container-fluid">
+      <a class="navbar-brand" href="/jeux">Games</a>
+
+      <div class="collapse navbar-collapse" id="navbarSupportedContent">
+        <ul class="navbar-nav me-auto mb-2 mb-lg-0">
+          <li class="nav-item">
+            <a class="nav-link active" aria-current="page" href="/jeux">Home</a>
+          </li>
+          <li class="nav-item">
+            <a class="nav-link" href="/jeux/new-game.php">Ajouter un jeu</a>
+          </li>
+
+
+        </ul>
+        <ul class="mr-5">
+          <li>
+            <?php
+            if (isset($_SESSION['isAuth'])) {
+
+              echo '
+                <li class="nav-item dropdown">
+                <a class="nav-link dropdown-toggle" href="#" role="button" data-bs-toggle="dropdown" aria-expanded="false">
+                  Bonjour '.$_SESSION['isAuth'].'
+                </a>
+                <ul class="dropdown-menu">
+                  <li><a class="dropdown-item" href="#">Profil</a></li>
+                  <li><hr class="dropdown-divider"></li>
+                  <li><a class="dropdown-item" href="/jeux/user/logout.php">Déconnexion</a></li>
+                </ul>
+                </li>
+';
+            } else {
+              echo '
+            <ul class="navbar-nav me-auto mb-2 mb-lg-0">
+              <li class="nav-item">
+                 <a class="nav-link" href="/jeux/user/register.php">S\'enregistrer |</a>
+              </li>
+              <li class="nav-item">
+                <a class="nav-link" href="/jeux/user/login.php">Se Connecter</a>
+              </li>
+          </ul>
+            
+            ';
+            }
+
+
+            ?> </li>
+
+
+        </ul>
+
+      </div>
     </div>
-  </div>
-</nav>
+  </nav>
 
   <h1 class="text-center m-5">List des jeux! </h1>
   <div class="container">
@@ -55,7 +92,7 @@ include './utils/db.php';
 
         <?php
 
-        //  $sql = "SELECT * FROM jeux_video WHERE prix Between 10 AND 20";
+        //  $sql = "SELECT * FROM jeux_video WHERE ID Between 11 AND 20";
         //  $sql = "SELECT * FROM jeux_video ORDER BY prix DESC LIMIT 3";.
 
         $sql = "SELECT * FROM jeux_video LIMIT 3";
@@ -64,27 +101,39 @@ include './utils/db.php';
 
         if ($result->num_rows > 0) {
           while ($row = $result->fetch_assoc()) {
+          
             echo '
-    <tr>
-    <th scope="row">' . $row["ID"] . '</th>
-    <td>' . $row["nom"] . '</td>
-    <td>' . $row["possesseur"] . '</td>
-    <td>' . $row["prix"] . '</td>
-    <td>' . $row["console"] . '</td>
-    <td>' . $row["nbre_joueurs_max"] . '</td>
-    <td>' . $row["commentaires"] . '</td>
-    <td>
-      <a href="delete-game.php/?id='.$row["ID"].'">
-         <i class="fa-solid fa-trash"></i>
-      </a>
-      <a href="update-game.php/?id='.$row["ID"].'">
-         <i class="fa-solid fa-pen"></i>
-      </a>
-     
+                <tr>
+                <th scope="row">' . $row["ID"] . '</th>
+                <td>' . $row["nom"] . '</td>
+                <td>' . $row["possesseur"] . '</td>
+                <td>' . $row["prix"] . '</td>
+                <td>' . $row["console"] . '</td>
+                <td>' . $row["nbre_joueurs_max"] . '</td>
+                <td>' . $row["commentaires"] . '</td>
+                <td>
+                ';
+            if (isset($_SESSION['isAuth'])) {
+              echo '
+                  <a href="delete-game.php/?id=' . $row["ID"] . '">
+                  <i class="fa-solid fa-trash"></i>
+                  </a>
+                  <a href="update-game.php/?id=' . $row["ID"] . '">
+                    <i class="fa-solid fa-pen"></i>
+                  </a>
+                  ';
+            } else {
+              echo '';
+            }
 
-      <a href="show-one-game.php/?id='.$row["ID"].'" class="btn btn-primary"> Details...</a></td>
-  </tr>
-    ';
+            echo '
+
+                    <a href="show-one-game.php/?id=' . $row["ID"] . '" class="btn btn-primary">
+                      Details...
+                    </a>
+                </td>
+                </tr>
+            ';
           }
         } else {
           echo "0 results";
